requirements.txt.txt
pip uninstall py-cord
PyNaC1==1.3.0
dnspython==1.60.0
async-timeout==3.0.1

aiohttp==3.7.4.post0
aiosqlite==0.17.0
asn1crypto==1.5.1
asyncpg==0.25.0
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
dblpy==0.4.0
discord-buttons==0.5
discord-components==2.1.2
discord.py==2.0.0a3575+g45d498c1
discord-py-interaction==s 4.1.0
discord-ui=5.1.6
Flask==2.0.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
npm==0.1.1

orjson==3.6.7
pg==0.1
pg8000==1.24.1
Pillow==8.4.0
pip==22.0.4

psycopg2-binary==2.9.3
psycopg2-pool==1.1

PyOpenG==3.1.5
python-dateutil==1.5
ratelimiter==1.2.0.post0
requests==2.26.0
scramp==1.4.1
setuptools==58.1.0
topggpy==1.4.0
typing_extensions==4.0.0
urllib3==1.26.7
websockets==3.4
Werkzeug==2.0.2
yarl==1.7.2